feat

+++ app/src/app/modules/templates/shoes-template/mx-size.pipe.ts
@@ -5,7 +5,7 @@ import { Pipe, PipeTransform } from '@angular/core';
 })
 export class MxSizePipe implements PipeTransform {
   transform(value: string): unknown {
-    const us = parseInt(value);
+    const us = parseFloat(value);
     const mx = us - 2;
 
     return mx.toString();
